Add Noop DNS method ipv6 disabling to handle non-routing scenarios
Introduced a NoopDnsMethod to prevent DNS manipulation when routing is disabled. IPV6 will also not be disabled when not routing. The assumption is that the user should handle all security considerations at this point when routing is disabled. Firewall should also be disabled if routing is disabled, but that is a future problem.